+++ database/update_vacancies.sql
@@ -1,7 +1,7 @@
 -- used to update existing vacancies
 CREATE OR REPLACE PROCEDURE work_scraper.update_vacancies(
     vacancy_id INTEGER[], title TEXT[], employer TEXT[],
-    salary_min DOUBLE[], salary_max DOUBLE[],
+    salary_min DOUBLE PRECISION[], salary_max DOUBLE PRECISION[],
     is_hourly BOOLEAN[], remote BOOLEAN[],
     published TIMESTAMP[], expires TIMESTAMP[],
     country_code VARCHAR(8)[], city_name TEXT[],
